# specs/003-dashboard-fix/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dashboard Application API
  description: API for the dashboard application with task management, projects, tags, and notifications
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourapp.com
    description: Production server
paths:
...
  /api/v1/tags:
    get:
      summary: Get all user tags
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of user tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create a new tag
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '200':
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
  /api/v1/tags/{tag_id}:
    put:
      summary: Update a tag
      security:
        - BearerAuth: []
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
        '404':
          description: Tag not found
    delete:
      summary: Delete a tag
      security:
        - BearerAuth: []
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tag deleted successfully
        '404':
          description: Tag not found
  /api/v1/notifications:
    get:
      summary: Get all user notifications
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of user notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
    post:
      summary: Create a new notification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreate'
      responses:
        '200':
          description: Notification created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
  /api/v1/notifications/{id}/read:
    patch:
      summary: Mark notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 3
          maxLength: 100
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - userId
    TagCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        userId:
          type: string
      required:
        - name
        - userId
    TagUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    Notification:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          maxLength: 500
        type:
          type: string
          enum: [info, warning, success, error]
        read:
          type: boolean
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - userId
    NotificationCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          maxLength: 500
        type:
          type: string
          enum: [info, warning, success, error]
        userId:
          type: string
      required:
        - title
        - type
        - userId
...